Generate default LL server config and attach it to GH release (#6430)
