+++ redbot/cogs/audio/managed_node/__init__.py
@@ -1,0 +1,16 @@
+# Note: contents of this package are meant to be self-contained
+# and should not depend on anything in Red or on external dependencies.
+
+from .ll_server_config import generate_server_config, get_default_server_config
+from .ll_version import LAVALINK_BUILD_LINE, LavalinkOldVersion, LavalinkVersion
+from .version_pins import JAR_VERSION, YT_PLUGIN_VERSION
+
+__all__ = (
+    "generate_server_config",
+    "get_default_server_config",
+    "LAVALINK_BUILD_LINE",
+    "LavalinkOldVersion",
+    "LavalinkVersion",
+    "JAR_VERSION",
+    "YT_PLUGIN_VERSION",
+)
